+++ src/ItemsDialog.js
@@ -79,7 +79,7 @@ const ItemsDialog = ({
   onRowClick = noop,
   mutator,
   skippedItemId,
-  title,
+  title = '',
   instanceId,
 }) => {
   const [areItemsBeingLoaded, setAreItemsBeingLoaded] = useState(false);
@@ -231,10 +231,6 @@ ItemsDialog.manifest = {
     accumulate: true,
     fetch: false,
   },
-};
-
-ItemsDialog.defaultProps = {
-  title: '',
 };
 
 ItemsDialog.propTypes = {